+++ src/oihana/commands/exceptions/ExitException.php
@@ -3,7 +3,6 @@
 namespace oihana\commands\exceptions ;
 
 use Exception;
-use oihana\exceptions\ExceptionTrait;
 
 /**
  * Invoked when the user won't exist a command.
@@ -14,5 +13,4 @@ use oihana\exceptions\ExceptionTrait;
  */
 class ExitException extends Exception
 {
-    use ExceptionTrait ;
 }