+++ src/oihana/commands/options/ChownOption.php
@@ -2,6 +2,7 @@
 
 namespace oihana\commands\options;
 
+use oihana\options\Option;
 use function oihana\core\strings\hyphenate;
 use function oihana\files\isMac;