+++ src/oihana/commands/options/CommandOption.php
@@ -2,6 +2,7 @@
 
 namespace oihana\commands\options;
 
+use oihana\options\Option;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputOption;