+++ src/oihana/commands/options/ServerOption.php
@@ -2,6 +2,8 @@
 
 namespace oihana\commands\options;
 
+use oihana\options\Option;
+
 /**
  * The enumeration of the global server options.
  *