+++ src/oihana/commands/options/ServerOptions.php
@@ -3,6 +3,7 @@
 namespace oihana\commands\options;
 
 use oihana\enums\Char;
+use oihana\options\Options;
 
 /**
  * Represents server-related configuration options for a website.