+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+- oihana\commands\exceptions\MissingPassphraseException class
+
 ## [1.0.3] - 2025-08-13
 
 ### Added